refactor: split backend core modules
