+++ docs/dev-logs/2026-06-06-backend-core-split-bundle.md
@@ -1,0 +1,19 @@
+# Backend Core Split Bundle
+
+Date: 2026-06-06
+
+## Summary
+- Split the backend route/service cores into smaller registration modules and lifecycle wrappers.
+- Kept all public entrypoints as thin barrels to preserve imports.
+- Preserved existing behavior and verified the full test suite.
+
+## Highlights
+- `media-core` split into admin/public route modules.
+- `ai-core` split into analysis/generation route modules.
+- `shortform-core` split into processing/library route modules.
+- `media-pipeline-actions-core` split into request/callback action modules.
+- `learning-store-core` split into state/persistence wrappers.
+
+## Verification
+- `npm run verify` passed
+- `npm run test:backend` passed
